Extract telemetry from route controller
